Pass command safely to shell
This commit fixes a bug in the code which was triggered by file names
with spaces in them. Instead of mucking about with escape codes, the
solution used here is to break up the parameters into elements of an
array of strings and pass that array to popen. Each element is then
appropriately handled by ruby when actually executing the command.